[java] factorize code between Bincode and LCS
